# src/openapi/conciliacion-openapi.yaml
openapi: 3.0.3
info:
  title: Conciliación — API combinada (conciliacion + movimientos-api)
  version: "1.0.0"
  description: |
    Especificación **machine-readable** generada por ingeniería inversa del código fuente
    (no existía documentación previa ni anotaciones Scalar/OpenAPI en el repo).

    Cubre dos servicios hermanos:

    1. **conciliacion** — Next.js 14 (App Router). Rutas bajo `src/app/api/**/route.ts`.
       Autenticación por cookie de sesión JWT (`conciliacion_session`), con un segundo nivel
       de autorización por "acción" (permisos por perfil, ver `src/lib/perms.ts`).
    2. **movimientos-api** — Express + mssql (`movimientos-api/src/server.js`). Expone 4
       feeds de solo lectura que `conciliacion` sincroniza periódicamente (Tesorería/Dynatech,
       Egresos, Movimientos internos de caja, patas TBK de POS). Auth opcional por API key.

    ### Notas de fidelidad al código
    - Los `requestBody` se derivan de los `z.object(...)` (Zod) definidos inline en cada
      `route.ts` (no se encontraron schemas Zod compartidos en una carpeta `lib/validations`;
      cada ruta define los suyos localmente).
    - Las respuestas se infieren de los objetos armados antes de `NextResponse.json(...)` y,
      cuando aplica, de los `select`/`include` de Prisma. Los campos cuya forma exacta depende
      de un helper no leído en detalle (p.ej. `lib/dashboard/queries.ts`, `lib/consolidados/match.ts`)
...
      (Prisma `Decimal`/`BigInt` serializados), no como `number`.
...

...
servers:
...
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: conciliacion_session
      description: >-
        JWT (HS256, firmado con JWT_SECRET, expira a los 7 días) seteado por
        POST /api/auth/login. httpOnly, sameSite=lax. `getSession()` (src/lib/auth.ts) lo
        decodifica; muchas rutas de mutación además exigen un permiso ("acción") concreto
        vía `requireAccion`/`denyUnless` (src/lib/perms.ts) — ver el campo `x-accion-requerida`
        de cada operación.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Opcional en movimientos-api: solo se exige si la variable de entorno API_KEY está
        seteada en el servidor (config.js). Si no está seteada, el middleware deja pasar
        cualquier request (excepto que igual se recomienda configurarla en producción).
  parameters:
    FromDateParam:
      name: from
      in: query
      required: false
      schema: { type: string, format: date, example: "2026-08-01" }
      description: Fecha de inicio del rango (YYYY-MM-DD). Default habitual = inicio del mes actual.
    ToDateParam:
      name: to
      in: query
      required: false
      schema: { type: string, format: date, example: "2026-08-03" }
      description: Fecha de fin del rango (YYYY-MM-DD, inclusive). Default habitual = hoy.
...
    AccountIdParam:
      name: accountId
      in: query
      required: false
      schema: { type: string, format: uuid }
...
    QParam:
      name: q
      in: query
      required: false
      schema: { type: string }
      description: Búsqueda libre (case-insensitive) sobre los campos de texto relevantes del recurso.
  responses:
    Unauthorized:
      description: No autenticado (falta la cookie de sesión o el JWT es inválido/expiró).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { error: "No autorizado" }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail:
          description: Detalle opcional (p.ej. issues de Zod) — presente solo en algunos 400.
          nullable: true
...
    IsoDateTime:
      type: string
      format: date-time
...
paths:
...
  /api/consolidados/egresos-terceros:
    get:
      tags: [consolidados-egresos-terceros]
      summary: Listar egresos bancarios (OUT) que no matchean ninguna entidad interna
      parameters:
        - { $ref: "#/components/parameters/FromDateParam" }
        - { $ref: "#/components/parameters/ToDateParam" }
        - { $ref: "#/components/parameters/AccountIdParam" }
        - { $ref: "#/components/parameters/QParam" }
        - { name: quality, in: query, schema: { type: string, enum: [con_rut, solo_nombre, sin_info] } }
      responses:
        "200":
          description: >-
            Excluye cuentas de uso parcial, descartados, comisiones/dif menor y los
            movimientos ya resueltos por asiento manual (estado GENERADO o EMITIDO).
          content:
            application/json:
              schema:
                type: object
                properties:
                  from: { $ref: "#/components/schemas/IsoDateTime" }
                  to: { $ref: "#/components/schemas/IsoDateTime" }
                  rows: { type: array, items: { type: object } }
                  totals: { type: object }
                  qualityCount: { type: object }
                  facets: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
...